+++ CHANGELOG.md
@@ -22,6 +22,7 @@ You can also check the
   - Clicking on Y axis title in bar charts now open Data tab in the details
     panels, as in other chart types
   - Fixed parsing problems with API routes
+  - Fixed GitLab CI/CD pipeline
 
 ### 6.1.2 - 2025-10-22